Add files via upload
RecorderAndExtract.R is an R script that records the human voice and does feature extraction on any .wav files in the folder that is passed through the processFolder function. SVMSound is a Support Vector Machine Model that classifies the gender of the person speaking. The sound_svm.rda is the saved model that was trained on 75% of the original data.